Flag collors mapper
