+++ EU4toV2/Data_Files/Eu4ToVic2DefaultConfiguration.xml
@@ -176,7 +176,7 @@
 	  <preference>
             <name>randomiseRgos</name>
             <friendlyName>Randomise RGOs?</friendlyName>
-            <description>If set, RGOs will be shuffled within similar culture and terrain provinces.</description>
+            <description>If set, RGOs will be shuffled within similar climate and terrain provinces.</description>
             <entryOptions>
               <entryOption>
                 <name>no</name>
