+++ EU4toV2/EU4ToV2.vcxproj.filters
@@ -33,9 +33,6 @@
     <ClCompile Include="Source\EU4World\EU4Version.cpp">
       <Filter>EU4World</Filter>
     </ClCompile>
-    <ClCompile Include="Source\Mappers\FlagColorMapper.cpp">
-      <Filter>Mappers</Filter>
-    </ClCompile>
     <ClCompile Include="Source\Mappers\GovermentMapper.cpp">
       <Filter>Mappers</Filter>
     </ClCompile>
@@ -438,6 +435,9 @@
     <ClCompile Include="Source\Mappers\CultureMapper\CultureMappingRule.cpp">
       <Filter>Mappers\CultureMapper</Filter>
     </ClCompile>
+    <ClCompile Include="Source\Mappers\FlagColors\FlagColorMapper.cpp">
+      <Filter>Mappers\FlagColors</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Source\Configuration.h" />
@@ -467,9 +467,6 @@
     </ClInclude>
     <ClInclude Include="Source\EU4World\EU4Version.h">
       <Filter>EU4World</Filter>
-    </ClInclude>
-    <ClInclude Include="Source\Mappers\FlagColorMapper.h">
-      <Filter>Mappers</Filter>
     </ClInclude>
     <ClInclude Include="Source\Mappers\GovernmentMapper.h">
       <Filter>Mappers</Filter>
@@ -879,6 +876,9 @@
     <ClInclude Include="Source\Mappers\CultureMapper\CultureMappingRule.h">
       <Filter>Mappers\CultureMapper</Filter>
     </ClInclude>
+    <ClInclude Include="Source\Mappers\FlagColors\FlagColorMapper.h">
+      <Filter>Mappers\FlagColors</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Filter Include="EU4World">
@@ -988,6 +988,9 @@
     </Filter>
     <Filter Include="Mappers\CultureMapper">
       <UniqueIdentifier>{0d5ba2f0-440e-4b34-91ef-d1bb99adcae2}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Mappers\FlagColors">
+      <UniqueIdentifier>{fc93dde8-f3e2-4484-9ad6-8374d24d3f5d}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
 </Project>
